@@ -1,9 +1,9 @@
 name: anim_search_bar
 description: A flutter package that has an animated search bar with loads of customization
-version: 1.0.1
+version: 2.0.0
 homepage: https://github.com/Imgkl/anim_search_bar
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
   flutter: ">=1.17.0"
 
 dependencies:
